% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xopt-package.R
\docType{package}
\name{xopt-package}
\alias{xopt}
\alias{xopt-package}
\title{xopt: State-of-the-Art Optimization with Automatic Differentiation}
\description{
Modern optimization framework combining quasi-Newton BFGS (ucminfcpp),
automatic differentiation (XAD), and zero-copy tensor operations (xtensor-r).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alrobles/xopt}
  \item Report bugs at \url{https://github.com/alrobles/xopt/issues}
}

}
\author{
\strong{Maintainer}: Angel Robles \email{angel.robles@example.com}

}
