% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_tracer.R
\name{xopt_ad_trace}
\alias{xopt_ad_trace}
\title{Create an AD-traced version of an R objective closure}
\usage{
xopt_ad_trace(fn)
}
\arguments{
\item{fn}{Objective function closure.}
}
\value{
A function with traced environment.
}
\description{
Builds a masked evaluation environment that maps selected base/stats
functions to xad-r adjoint implementations when available.
}
