% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_tracer.R
\name{xopt_auto_gradient}
\alias{xopt_auto_gradient}
\title{Automatic gradient dispatch with optional R-side AD tracing}
\usage{
xopt_auto_gradient(fn, par, tracer = TRUE)
}
\arguments{
\item{fn}{Objective function.}

\item{par}{Parameter vector where gradient is computed.}

\item{tracer}{Logical; if TRUE apply function masking prior to AD.}
}
\value{
Numeric gradient vector.
}
\description{
Automatic gradient dispatch with optional R-side AD tracing
}
