% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xopt_chol_grad}
\alias{xopt_chol_grad}
\title{Adjoint gradient of sum(L) where L = chol(A)}
\usage{
xopt_chol_grad(A)
}
\arguments{
\item{A}{numeric SPD matrix}
}
\value{
numeric matrix of ∂(sum L) / ∂A
}
\description{
Adjoint gradient of sum(L) where L = chol(A)
}
