% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xopt_chol_impl}
\alias{xopt_chol_impl}
\title{Forward-mode Cholesky on double}
\usage{
xopt_chol_impl(A)
}
\arguments{
\item{A}{numeric SPD matrix (n x n)}
}
\value{
lower-triangular Cholesky factor L (n x n) with A = L L^T
}
\description{
Forward-mode Cholesky on double
}
