% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{xopt_control}
\alias{xopt_control}
\title{Optimization control parameters}
\usage{
xopt_control(
  grtol = 1e-06,
  xtol = 1e-12,
  maxiter = 500,
  trace = FALSE,
  stepmax = 1,
  ftol = 1e-12
)
}
\arguments{
\item{grtol}{Gradient tolerance. Optimization stops when max|gradient| <= grtol.}

\item{xtol}{Step tolerance. Optimization stops when step length <= xtol.}

\item{maxiter}{Maximum number of function evaluations.}

\item{trace}{Logical; if TRUE, print trace output during optimization.}

\item{stepmax}{Initial trust region radius.}

\item{ftol}{Relative objective-function tolerance.}
}
\value{
A list of control parameters.
}
\description{
Create a list of control parameters for xopt optimization.
}
