% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xopt_inv_grad}
\alias{xopt_inv_grad}
\title{Adjoint gradient of sum(A^{-1})}
\usage{
xopt_inv_grad(A)
}
\arguments{
\item{A}{numeric SPD matrix}
}
\value{
numeric matrix of ∂ sum(A^{-1}) / ∂A
}
\description{
Adjoint gradient of sum(A^{-1})
}
