% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xopt_inv_impl}
\alias{xopt_inv_impl}
\title{Forward-mode SPD inverse on double}
\usage{
xopt_inv_impl(A)
}
\arguments{
\item{A}{numeric SPD matrix (n x n)}
}
\value{
numeric matrix A^{-1}
}
\description{
Forward-mode SPD inverse on double
}
