% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xopt_logdet_grad}
\alias{xopt_logdet_grad}
\title{Adjoint gradient of logdet(A)}
\usage{
xopt_logdet_grad(A)
}
\arguments{
\item{A}{numeric SPD matrix}
}
\value{
numeric matrix of ∂ log|A| / ∂A (expected A^{-1})
}
\description{
Adjoint gradient of logdet(A)
}
