% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xopt_logdet_impl}
\alias{xopt_logdet_impl}
\title{Forward-mode log-determinant on double}
\usage{
xopt_logdet_impl(A)
}
\arguments{
\item{A}{numeric SPD matrix (n x n)}
}
\value{
scalar log|A|
}
\description{
Forward-mode log-determinant on double
}
